Implemented $ref in base object and refactored create method
