+++ src/Objects/Info.php
@@ -47,15 +47,6 @@ class Info extends BaseObject
     protected $version;
 
     /**
-     * @param string|null $objectId
-     * @return \GoldSpecDigital\ObjectOrientedOAS\Objects\Info
-     */
-    public static function create(string $objectId = null): self
-    {
-        return new static($objectId);
-    }
-
-    /**
      * @param string|null $title
      * @return \GoldSpecDigital\ObjectOrientedOAS\Objects\Info
      */
@@ -136,7 +127,7 @@ class Info extends BaseObject
     /**
      * @return array
      */
-    public function toArray(): array
+    protected function generate(): array
     {
         return Arr::filter([
             'title' => $this->title,
