+++ src/Objects/Parameter.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace GoldSpecDigital\ObjectOrientedOAS\Objects;
 
 use GoldSpecDigital\ObjectOrientedOAS\Contracts\SchemaContract;
-use GoldSpecDigital\ObjectOrientedOAS\Traits\Referenceable;
 use GoldSpecDigital\ObjectOrientedOAS\Utilities\Arr;
 
 /**
@@ -25,8 +24,6 @@ use GoldSpecDigital\ObjectOrientedOAS\Utilities\Arr;
  */
 class Parameter extends BaseObject
 {
-    use Referenceable;
-
     const IN_QUERY = 'query';
     const IN_HEADER = 'header';
     const IN_PATH = 'path';
@@ -104,15 +101,6 @@ class Parameter extends BaseObject
      * @var \GoldSpecDigital\ObjectOrientedOAS\Objects\MediaType[]|null
      */
     protected $content;
-
-    /**
-     * @param string|null $objectId
-     * @return \GoldSpecDigital\ObjectOrientedOAS\Objects\Parameter
-     */
-    public static function create(string $objectId = null): self
-    {
-        return new static($objectId);
-    }
 
     /**
      * @param string|null $objectId
@@ -322,12 +310,8 @@ class Parameter extends BaseObject
     /**
      * @return array
      */
-    public function toArray(): array
+    protected function generate(): array
     {
-        if ($this->ref !== null) {
-            return ['$ref' => $this->ref];
-        }
-
         $examples = [];
         foreach ($this->examples ?? [] as $example) {
             $examples[$example->objectId] = $example->toArray();
