+++ src/Objects/SecurityRequirement.php
@@ -24,15 +24,6 @@ class SecurityRequirement extends BaseObject
     protected $scopes;
 
     /**
-     * @param string|null $objectId
-     * @return \GoldSpecDigital\ObjectOrientedOAS\Objects\SecurityRequirement
-     */
-    public static function create(string $objectId = null): self
-    {
-        return new static($objectId);
-    }
-
-    /**
      * @param \GoldSpecDigital\ObjectOrientedOAS\Objects\SecurityScheme|string|null $securityScheme
      * @throws \GoldSpecDigital\ObjectOrientedOAS\Exceptions\InvalidArgumentException
      * @return \GoldSpecDigital\ObjectOrientedOAS\Objects\SecurityRequirement
@@ -72,7 +63,7 @@ class SecurityRequirement extends BaseObject
     /**
      * @return array
      */
-    public function toArray(): array
+    protected function generate(): array
     {
         return Arr::filter([
             $this->securityScheme => $this->scopes ?? [],
